Set FixedExpenseCategory to limited categories

+++ src/test/java/team/easytravel/model/trip/BudgetTest.java
@@ -16,7 +16,7 @@ public class BudgetTest {
 
     @Test
     public void constructor_invalidPhone_throwsIllegalArgumentException() {
-        Double invalidBudget = 0.0;
+        Integer invalidBudget = 0;
         Assert.assertThrows(IllegalArgumentException.class, () -> new Budget(invalidBudget));
     }
 
@@ -26,13 +26,13 @@ public class BudgetTest {
         Assert.assertThrows(NullPointerException.class, () -> Budget.isValidBudget(null));
 
         // invalid phone numbers
-        assertFalse(Budget.isValidBudget(-000000.0)); // negative 0s
-        assertFalse(Budget.isValidBudget(-2.0)); // spaces only
-        assertFalse(Budget.isValidBudget(999999999.0)); // very big number
+        assertFalse(Budget.isValidBudget(-000000)); // negative 0s
+        assertFalse(Budget.isValidBudget(-2)); // spaces only
+        assertFalse(Budget.isValidBudget(999999999)); // very big number
 
         // valid phone numbers
-        assertTrue(Budget.isValidBudget(911.0)); // exactly 3 numbers
-        assertTrue(Budget.isValidBudget(000000001.0)); //extra 0s
+        assertTrue(Budget.isValidBudget(911)); // exactly 3 numbers
+        assertTrue(Budget.isValidBudget(000000001)); //extra 0s
 
     }
 }
